feat(core): allow using SNAPSHOT statements as an alternative to BACKUP
