+++ core/src/main/java/io/questdb/cairo/O3Utils.java
@@ -46,7 +46,11 @@ public class O3Utils {
     ) throws SqlException {
         final MessageBus messageBus = cairoEngine.getMessageBus();
         final int workerCount = workerPool.getWorkerCount();
-        final O3PartitionPurgeJob purgeDiscoveryJob = new O3PartitionPurgeJob(messageBus, workerPool.getWorkerCount());
+        final O3PartitionPurgeJob purgeDiscoveryJob = new O3PartitionPurgeJob(
+                messageBus,
+                cairoEngine.getSnapshotAgent(),
+                workerPool.getWorkerCount()
+        );
         workerPool.assign(purgeDiscoveryJob);
 
         // ColumnPurgeJob has expensive init (it creates a table), disable it in some tests.
