+++ core/src/main/java/io/questdb/cutlass/text/CopyTask.java
@@ -493,6 +493,7 @@ public class CopyTask {
                             DefaultLifecycleManager.INSTANCE,
                             root,
                             cairoEngine.getDdlListener(tableToken),
+                            NoOpDatabaseSnapshotAgent.INSTANCE,
                             cairoEngine.getMetrics()
                     )
             ) {
@@ -875,6 +876,7 @@ public class CopyTask {
                             DefaultLifecycleManager.INSTANCE,
                             importRoot,
                             engine.getDdlListener(tableToken),
+                            NoOpDatabaseSnapshotAgent.INSTANCE,
                             engine.getMetrics()
                     )
             ) {
